refactor: 최상위 예외클래스 패키지 이동
