+++ README.md
@@ -487,11 +487,6 @@ pobi:--
     - [x] 진행 중인 게임 확인
     - [x] 종료된 게임 확인
 
-
-- [ ] 승자 확인하기
-    - [ ] 단독 승자 확인
-    - [ ] 공동 승자 확인
-
 <h2>2. 이동 정책(MovementPolicy) 유스케이스</h2>
 
 <h4>
