--- src/main/java/racingcar/common/exception/RacingCarException.java
@@ -1,7 +1,0 @@
-package racingcar.common.exception;
-
-public class RacingCarException extends RuntimeException{
-    public RacingCarException(String message){
-        super(message);
-    }
-}
