+++ src/main/java/racingcar/exception/RacingCarException.java
@@ -1,0 +1,7 @@
+package racingcar.exception;
+
+public class RacingCarException extends RuntimeException {
+    public RacingCarException(String message) {
+        super(message);
+    }
+}
